Time: 88 ms (83.35%), Space: 76.9 MB (22.4%) - LeetHub

+++ 1894-find-the-student-that-will-replace-the-chalk/1894-find-the-student-that-will-replace-the-chalk.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int chalkReplacer(vector<int>& chalk, int k) {
         int num=0;
-        long long int sum=0;
+        long int sum=0;
         int i=0;
         for(i=0;i<chalk.size();i++)
         {
@@ -18,6 +18,5 @@ public:
             }
         }
         return i;
-        
     }
 };
